Fintech-Insights-main/requirements.txt
Flask
requests
python-dotenv
psycopg2-binary
google-generativeai
